Move scale notes to store
